IterativePower

+++ IterativePower.go
@@ -5,7 +5,7 @@ import "math"
 func IterativePower(nb int, power int) int {
 	var result int
 
-	c := math.Pow(nb, power)
+	c := math.Pow(float64(nb), float64(power))
 
 	if power < 0 {
 		return 0
